requirements.txt
# For MongoDB
dnspython
# Discord
discord.py
# Voice activity
PyNaCl
# Get mp3 link for video
youtube_dl
# Search for youtube videos
youtube_search
# Parse title of youtube video
youtube_title_parse
# Get lyrics
lyricsgenius
# MongoDB
PyMongo
# To convert emoji to :emoji_name:
emoji
# Check for NSFW content
nudenet
# Pafy to get thumbnail of video
pafy
# Download thumbnail
requests